+++ frontend/javascripts/messages.js
@@ -170,6 +170,7 @@ In order to restore the current window, a reload is necessary.`,
   "annotation.was_re_opened": "Annotation was reopened",
   "annotation.delete": "Do you really want to reset and cancel this annotation?",
   "annotation.was_edited": "Successfully updated annotation",
+  "annotation.shared_teams_edited": "Successfully updated the shared teams for the annotation",
   "project.delete": "Do you really want to delete this project?",
   "project.increase_instances":
     "Do you really want to add one additional instance to all tasks of this project?",